requirements.txt
langchain>=0.2.11
langgraph>=0.2.21
langchain-openai>=0.1.22
langchain-mcp-adapters>=0.0.15
fastmcp>=0.4.1
httpx>=0.27.0
python-dotenv>=1.0.1
